+++ src/utils.js
@@ -75,8 +75,11 @@ function requestLogger(req) {
         req.requestId = uuidv4();
     }
     if (req.method === 'POST') {
-        // Ensure we don't log the token
-        const loggedBody = Object.assign({}, req.body, {token: '<redacted>'});
+        let loggedBody = Object.assign({}, req.body);
+        if (loggedBody.token) {
+            // Ensure we don't log the token
+            loggedBody.token = '<redacted>';
+        }
         logger.log('info', `${req.method} ${req.path}: ${tryStringify(loggedBody)}`, {requestId: req.requestId});
     } else {
         logger.log('info', `${req.method} ${req.path}`,{requestId: req.requestId});